requirements.txt
streamlit
anthropic
pillow
pytesseract
ics
pytz
python-dateutil